requirements.txt
pandas
numpy
requests
python-dotenv
tqdm
yfinance
beautifulsoup4
schedule
scikit-learn
pyarrow
pytz
